+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.3.0](#030)
   - [0.2.1](#021)
   - [0.2.0](#020)
+
+## 0.3.0
+
+04/03/2024
+
+- Feat: added the possibility to specify the wallet provider in the context with the `provider` property
+- Feat: implemented the Dfinity wallet provider, using the internet computer identity
 
 ## 0.2.1